packages/ingestion/pyproject.toml
[project]
name = "tanmay-ingestion"
version = "0.1.0"
description = "Create with Tanmay, ingestion + processing pipeline"
requires-python = ">=3.12"
dependencies = [
    "celery[redis]>=5.4.0",
    "redis>=5.2.0",
    "pydantic>=2.9.0",
    "pyyaml>=6.0.2",
    "httpx>=0.27.0",
    "structlog>=24.4.0",
    "yt-dlp>=2024.10.22",
    "youtube-transcript-api>=0.6.3",
    "feedparser>=6.0.11",
    "apify-client>=1.8.0",
    "deepgram-sdk>=3.7.0",
    "voyageai>=0.3.0",
    "anthropic>=0.39.0",
    "qdrant-client>=1.12.0",
    "tiktoken>=0.8.0",
    "boto3>=1.35.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.3.0", "ruff>=0.7.0", "mypy>=1.13.0"]

[tool.ruff]
line-length = 120
target-version = "py312"
